fix spacing, cleanup code
